Backend/sid_text2sql/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
psycopg2-binary>=2.9.9
openai>=1.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
streamlit>=1.28.0

